Move reset tests to Robolectric to avoid state setup problems
